Move integration tests to file-lock-manager tests
